2015 Behavioral Risk Factor Surveillance System

A subset of data from the Centers for Disease Control and Prevention 2015 Behavioral Risk Factor Surveillance System (BRFSS) Survey

Usage

brfss

Format

brfss

A data frame with 5,000 rows and 17 columns:

COPD

Chronic obstructive pulmonary disease

smoke

Smoke

sex

Sex

weight

Weight

average_drinks

Average drinks answers to: during the past 30 days, when you drank, how many drinks did you drink on average?

race_white, race_black, race_hispanic, race_other

Race group

age_18to24, age_25to34, age_35to44, age_45to54, age_55to64, age_over65

Age groups

Source

http://static.lib.virginia.edu/statlab/materials/data/brfss_2015_sample.csv